test: move fs fakes to a common file
This will be useful for other tests shortly.
